pyproject.toml
[project]
name = "map-web-ui"
version = "0.1.0"
description = "Web client for managing JAIRO Cloud groups."
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "celery>=5.5.3",
    "flask>=3.1.2",
    "flask-sqlalchemy>=3.1.1",
    "psycopg[binary]>=3.2.10",
    "pydantic-settings>=2.11.0",
    "pydantic[email]>=2.12.2",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "sqlalchemy-utils>=0.42.0",
]

[dependency-groups]
dev = [
    "celery-types>=0.23.0",
    "pyright>=1.1.406",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "taskipy>=1.14.1",
]

[tool.setuptools.packages.find]
include = ["server"]


[tool.taskipy.tasks]
# refer to https://github.com/taskipy/taskipy?tab=readme-ov-file#usage
test = "pytest"
"test:cov" = "pytest --cov=server --cov-report=term --cov-report=html"
lint = "uvx ruff check"
"lint:fix" = "uvx ruff check --fix"
format = "uvx ruff format"
typecheck = "pyright"
check-updates = "uvx --from=pip-check-updates pcu pyproject.toml"

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = [
    "--import-mode=importlib",
]

[tool.coverage.run]
source = ["server"]
omit = ["*/__init__.py", "*/tests/*"]

[tool.ruff.lint]
# refer to https://docs.astral.sh/ruff/settings/#lint
ignore = [
  "D203",
  "D213",
  "TC006",
  "COM812",
  "ANN201",
  "TRY400",
]
preview = true
select = ["ALL"]

[tool.ruff.per-file-ignores]
"tests/*" = [
  "D",
  "S101",
  "S106",
  "ANN",
  "INP001",
  "ARG001",
  "TC006",
  "DTZ005",
  "PLC2701",
  "SLF001",
  "SIM300",
]

[tool.ruff.lint.isort]
# refer to https://docs.astral.sh/ruff/settings/#lintisort
combine-as-imports = true
known-first-party = ["server"]
lines-between-types = 1

[tool.ruff.analyze]
# refer to https://docs.astral.sh/ruff/settings/#analyze
detect-string-imports = true

[tool.ruff.format]
# refer to https://docs.astral.sh/ruff/settings/#format
docstring-code-format = true
docstring-code-line-length = 80
line-ending = "lf"
preview = true
skip-magic-trailing-comma = false

[tool.pyright]
extraPaths = ["server"]
typeCheckingMode =  "standard"
